Main procedure in intermediate well-formedness
Have the Intermediate machine follow exactly the same conventions as the
Source language: namely, stating that if a program has the main
component in its interface it is required to declare a main procedure.
The encoding remains more complex than at the Source due to differences
in encoding, but otherwise few changes are needed and no existing proofs
break beyond easy repair.

It is interesting to note that the old form of the Intermediate
condition is exactly the contrapositive of the Source condition (and the
modified Intermediate condition).