Strengthen wfprog_main_existence in Source
Consider a notion of well-formed program in which the existence of a
main procedure and the presence of the main component in the program
interface are inextricably linked. This is opposed to the currently
standard view, in which presence in the interface implies the existence
of the main procedure. (But if the main component is not in the
interface, there could be a main procedure or not, and provided by
someone else, which does not make much sense.)

The change invoves some small modifications to existing proofs.