Add closed_program to MultiSemantics
As in previous sections, we add the hypothesis that the program (the
linking of p and c) must be closed. In principle, this restricts the
situations in which its simulations can be applied, but in practice has
no effect in our development. Moreover, it is not much of a stretch
given the rather strong implications of PS.mergeable_interfaces in its
present form.