High-level proof of main prefix composition result
With a little thought, this property can be expressed as a simple
combination of three other, natural results, each of which should have a
relatively simple proof, certainly simpler than it would be in
combination with the others.

In the final proof, there are three cases based on the kind of finite
trace prefix. The "proper prefix" case follows from the posed lemmas,
whereas the other two correspond to the existing result
partial_program_composition (for termination) and its homologue for
error during execution.