Port some results from Source.PS to Intermediate
As expected, not only are Source statements the same on the Intermediate
level, but their proofs are also reusable almost verbatim; where
differences occur, they have been due to minor, trivial discrepancies
between source and intermediate levels.

Of course, this points to the extraction of a common substrate
underlying both partial semantics.