Simplify memory merges where possible
The two cases changed here are different from legitimate applications of
the more complex memory partition lemmas not only on two interfaces, but
on two memories as well. This change allows us to complete the statement
of the lemmas that cover the latter case based on the contexts where
they are used.