Complete statements of partial memory merges
The lemmas are enriched with some additional premises in the form that
the exist in the proofs where they are used that allow us to determine
that the two domains of the memories are exactly those of the two
interfaces, without any external "garbage". The proper thing to do is to
have a lemma that uses the domains directly, and based on which the two
proofs of the more concrete lemmas are formulated, or a similar setup.

For this reason the difficulty rating of the lemmas is bumped up a
little, but their proofs are not to be surprising in any way.