Decompose specialized reasoning of memory merges
The specialized reasoning of to_partial_memory_merge_memories_* can be
seen (among other possibilities) as a quest to determine that the
domains of the two memories coincide with the domain of the two
interfaces under consideration. We can move from the hypotheses in our
proofs of simulation to that simpler assertion by a couple of simpler
results, and this transformation can further simplify either the
statement of the memory merge theorems (by slightly complicating their
appearance in the main proof) or their proof.

Co-authored-by: Florian Groult <florian.groult@inria.fr>