Fix statement of does_prefix_star
Instead of formulating this lemma in terms of an arbitrary semantics,
use the CS.sem that is defined and used in this module. This allows us
to avoid dead ends in the proof by using facts such as the existence of
initial states. Other properties of the semantics, such as singleton
traces, also come in handy.