State threeway simulation on plain stars
Previous simulation results on runs of both program and context
simultaneously to obtain a simulation of the combination of the two have
followed the "initial" pattern of using "fancy stars" of various kinds.
The decorated, turn-taking stars, however, always depend on the same
number of steps being taken on all three runs.

As in other points in the development, a strong reliance of taking the
same number of steps has proven to make things more complicated and we
have had to work around it several times. It is very likely that, here
again, a more general result will be needed to prove our top-level
theorems easily, even if this can build on an indirect lockstep guided
by the number of steps, as already present in the Composition module. It
is also possible that this may simplify some of the intermediate results
that are already in place.