Refactor behavior_prefix_star
As usual, we will want to simplify section variables as much as possible
once all the proofs are finished.

Now it is possible to give proofs of both related subgoals in
partial_programs_composition_prefix, whereas previously only the first
one could be handled due to the misplacement of the lemma.