Existence of PS initial states
State the existence of initial states in the partial semantics given a
program and a context and their individual and joint well-formedness
conditions.

Use the contradictory form of this property to finish the proof of
partial_programs_composition_prefix, which was the higher-level admitted
composition result on which the top-level proof depended. From here on,
we will take on the remaining dependencies top-down, as [Print
Assumptions] observes them.