Refine and prove prog_main_same_interface
First, we slightly refine the statement that relates the mains of two
programs that share the same interface. We cannot exactly say that the
main identifier will be exactly the same, at least not in the abstract
view that we are taking. What we can say is that (given, of course, the
fact that the programs are well formed) they either have mains or they
do not. We rephrase the lemma into the mainless variant, as it is a
little more lightweight and in the situation where we are using this
result, namely to discharge contradictions, either one works.

Having done that, we adjust the use of the revised and proved lemma in
the proof of merged_initial_states.