Intermediate PS initial state determinism
Same as at the Source level, initial states are deterministic at the
intermediate level, also in the partial semantics. Here we give a result
that relies on two programs instead of a program and a context,
corresponding to the Composition environment where its use is needed,
however a simplified statement can be looked for.

The small details omitted from the proof constitute a small exercise in
refactoring reasoning patterns on partial memories and main procedures,
which may apply to similar reasoning in, say, merged_initial_states, and
possibly other places.