Complete behavior_prefix_star (axiom index 6)
Using initial state determinism, the proof of behavior_prefix_star is
finished. The final theorem does not depend on mergeability of the
interfaces, which is removed from both uses of the lemma.

With this, the number of axioms in the Composition module is down to
six. We will attack the missing proofs systematically to bring the
number down; the process may not be monotonic, as the number of simple
auxiliary results may yet increase, but the immediate goal is to bring
the number of interesting missing results in the module down to zero.