Finish initial_states_mergeability (axiom index 5)
The symmetric case completes the proof. Here the most lightweight way to
deal with the symmetries is not defining, say, lemmas that deal with
memory operations where the preserved part is respectively on the left
or on the right. Rather, we can apply commutativity of unions and use a
single instance of the results for both sides, as is done here.

After doing this, we are left with five admitted results in Composition,
and will resume our assault on the remaining ones.