Sample case in mergeable_states_step_CS
Model solution for one of the main cases of the lemma, on INop. Because
the nested inversion on the other CS step is not directly linked to the
first by the same program, this case analysis results in a number of
subgoals, each of which is used to compose the final, collective step
from the pieces of the two programs, each substituting one on the halves
of the original program. It will become increasingly important to
refactor the differences between sub-cases via tactics.