Selection lemma on imported procedures
This result is very similar to other related results that are found in
Intermediate.GlobalEnv and which simplify a program link by ascertaining
that the sought component is in one of the programs, therefore
discarding the other one in the lookup. imported_procedure operates on
interfaces and can be made independent of Source and Intermediate
levels by direct simplification.