Controlled rewrites and compositions
When dealing with procedure calls and returns, goals involve changes on
the stack whose simplification needs to be performed with care to avoid
excessive evaluation. A simple way to achieve this is stating rewrite
lemmas, trivial by computation but useful to stop after the desired
amount of evaluation has been performed.