mergeable_states_step_trans: base jump sub-goals
Provide specialized treatment to cover jump instructions, establishing
component equality and treating PC increments. This part differs from
the treatment of mergeable_states_step_CS, but is of a similar kind,
however the tactics are at the moment slightly more involved. It would
be interesting to refactor a bit more and simplify steps, if possible,
as more first-level sub-goals (resulting from the first inversion on a
CS step) are added.