mergeable_states_step_*: refactor stack rewrites
This pair of related lemmas exhibits blocks of common reasoning. Stack
rewrites (on a merged stack) need to be performed in several places in
mergeable_states_step_trans, with small variations, therefore we
encapsulate the common operations, originally given for its sibling,
mergeable_states_step_CS. The tactic is currently parametric on some
components of the context, but these are always known and should be
inferred later.