threeway_multisem_mergeable: proof modulo symmetry
The inductive step is finished by simple case analysis on the trace
produced by the step. The epsilon case is different than the trace
event-producing case, because it need not correspond to any number of
steps on the star side. One could relate two applications of the
previously posed lemma, mergeable_states_star_E0, on program and context
and cross-correlate their results, or proceed by case analysis on which
side has control, as done here.