threeway_multisem_star: proof outline
As one could expect, the proof schema is remarkably similar to that of
threeway_multisem_mergeable. Case analyzing on the trace produced by the
single step after the IH has been applied seems to be the most
convenient way to decompose the resulting sub-goals into a small number
of easier and reoccurring results.