Move threeway_multisem_mergeable
The common use of symmetries, previous results, etc., for which a fixed
set of section hypotheses holds is a pain, but I wish to avoid changing
the proofs or complicating the section-based theorem statements at this
point. We will add a small number of dedicated sections while this
situation is untangled.