threeway_multisem_step_E0: close proof by Ltac
Case analysis on the CS step to prove the CS sub-goal is given its own
Ltac block. In order to make Ltac happy, a few minor changes are made to
the existing proof script. Strangely, in two instances the standalone
script cannot detect names of hypotheses that were generated by previous
steps in the script, so here we resort to enriched [match goal]s when
needed. Otherwise, all the names above the line referenced by the script
as passed as arguments.