Symmetric partial memory merge lemmas
Permutations on existing lemmas that establish the domains of the parts
in slightly different, but equivalent ways as those lemmas already
posed. This is annoying and in part relies on execution provenance,
which by now appears to be stronger than strictly needed. The results
themselves are perfectly straightforward, but there is a good chance
that their proofs, and those of their siblings, can be made
substantially simpler by an appropriate weakening of the provenance
condition, and their (otherwise uninteresting) proofs should be deferred
until that possibility is explored.