Common inversion tactics
Both lemmas mergeable_states_step_trans and mergeable_states_step_CS
share much of their basic proof strategy by case analysis, arising from
identical sets of hypotheses. The same conditions apply to the sequence
of inversions in each lemma, including the discharge of contradictory
sub-goals and the required reordering of inversions with respect to the
proof sketches, which result in particular in two separate if symmetric
cases for each call and return instructions.