Multiple fixes to the definability which do not break the previous spec
Furthurmore, it helped getting rid of the NOP hack (which broke the
specs). Mainly, the changes consist in the use of pmap instead of map in
expr_of_trace, expr_of_event returns an option : no need to return dummy
expressions, only events triggered by the component return expressions.

The proof of well_formed_events_well_formed_program is fixed wrt this,
the counter value is defined by only the number of events triggered by a
component C, proofs up to 'definability_gen' are fixed and so is the
base case of the latter.