Rephrase top-level lemmas for recombination
The more abstract components of a proof of recombination (decomposition
+ composition) are modeled after the proof of recomposition, expected to
be a close relative of it. For now the same names are kept even as in
this new setting some should later find more apt designations.

Importantly, the notion of compatibility between states is no longer
given by the mergeability between two PS states (with all the
complications that entailed), but is stated in terms of a relation
between two CS states and a function that performs "state
recombination", both as yet unspecified. Together with these must come a
theorem relating them, i.e., nothing that whenever the relation is
satisfied by a pair of states, the function will succeed in producing a
result. We shall use option types to keep the obligations lightweight
while explicitly covering nonsensical conditions, and ruling these out
with the help of the relation.