Simplify interfaces in recombination
Opt for the less verbose formulation that does away with separate
interface variables and states their equalities with respect to each
pair of programs separately, phrasing these facts simply as interface
equalities and using one of those in theorem statements whenever needed.

There is an interesting technical advantage to doing this. If, within
the same section relying on the more verbose setup of variables and
hypotheses, the interface variables are simplified in the course of a
proof, application of lemmas from the same section relying on the
unified and vanished variables will fail.