Adapt composition_prefix as recombination_prefix
The topmost recomposition proof is, as expected, a simple variation on
the topmost composition proof (together with the proper proof it wraps,
in partial_program_compositions_prefix, owing to the absence of a PS-CS
wrapper). Both proofs play out in exactly the same way, with only small
differences owing to the phrasing of the theorems (because of the
cleaner conventions in state merging); several appeals to previously
separate simulations are already avoided.

Co-authored-by: Jérémy Thibault <jeremy.thibault@inria.fr>