Extensional, Composition-style mergeable states
Experiment with an "extensional" version of the mergeability relation
between pairs of states, a natural continuation of the current
definition in Composition. Moreover, to keep theorem statements
lightweight and as close to their Composition counterparts (and their
proofs) as possible while we do the portability, remove the option type
from the merge operation (still given as a placeholder) while ensuring
that the operation is only performed when the relation holds.