Initial scaffolding Recombination proofs
Following Composition and generalizing the mergeability relation to
expose interface information. It ends up being more verbose than its PS
counterpart, principally due to the involvement of two explicit pairs of
programs.

Sectioning and greedy inclusion of hypotheses in admitted lemmas make
some of the later proofs break until the now external hypotheses are
named explicitly, to be fixed when it is clear the high-level structure
continues to work, as expected.