Add an induction lemma for mergeable states
The lemma states the following: If 1) a property is true of two initial states (these states are always mergeable) 2) the same property is preserved by epsilon-step either on the right and on the left 3) the same property is preserved by doing the *same* non-silent step on both states

Then, the property is true for all mergeable states. I used this principle to prove one lemma, and I'm in the process of proving other lemmas too.