Rephrase (partial) state equality on epsilon star
The main lemma is recast in a closer form as it is present in PS, with
the original formulation being a simple corollary that will follow from
the definitions of state mergeability. (Adjustments in their uses need
to be finalized.)

Because we will be dealing with two pairs of programs and their (equal)
interfaces, some amount of rewriting is to be expected and could be
integrated into lemmas for common use.