Refactor stack recombination, add cons lemmas
Add a stack-based definition to the existing state-based one, and
reformulate the latter in terms of the former. This, in combination with
the rewrite lemmas on merged states, makes exposing the structure of the
recombined states easier.

In addition to those results, state lemmas to expose the top of a merged
stack based on program vs. context information, determining which state
to select from.