Complete/refactor recombination lockstep on call
The two cases, call and return, of the recombining lockstep
synchronization each are divided in two cases according to the location
of the target, program or context, which then selects which one of the
two program runs will be active until the next synchronization event.

In each of the two main cases the case analysis operates on a different
part of the context. The two resulting sub-goals are then almost
identical except for a few points where their complementary symmetry
comes into play. The parts which are homogeneous across sub-cases
differ across main cases, but it would be interesting to see if the
proof benefits from a unified treatment of both.