Complete/refactor recombination lockstep (return)
Remove the (trivial) admits left in the return case, moreover removing
their details from the proof script, and factor the case analysis in
this case so as to express common processing and symmetries.

Note, for example, that PC equalities generated by the nested inversion
on the two steps of the runs (in lockstep) are used in both cases.
Observe also that the main goal accepts an identical solution in both
cases. This and other observations can be used to simplify the proof
further.