Recombining definition of merge_memories
Memory partialization with respect to an interface filters the
components beloning to that interface from the memory. Therefore, if a
memory merge takes two memories (assumed compatible) and combines them
according to two interfaces (for program and interface), its result must
be the union of the program components (of the first interface) from the
first memory (i.e., partialized with respect to the context), and the
context components (of the second interface) from the second memory
(i.e., vice versa).