Prototype proof of recombining match_nostep
In PS-reliant composition, the match between stuck states comes as a
freebie -- after the significant proof effort required by proofs
involving the partial semantics. In this cleaner world, a bit of extra
work is required at this point.

The contradiction we seek arrives through a result that is closely
related to the recombining [threeway_multisem_step_E0], but covers all
kinds of CS steps (as opposed to only silent steps) and does so in a
slight rearrangement of the programs which makes application of that
result slightly cumbersome.

The result we are interested in is this: if the recombination of
[program_link p c'] can take a step, either [program_link p c] (if the
step takes place from p) or [program_link p' c'] (if the step takes
place in c') can take an equivalent step. We do not need to give an
expression of the end state of the modified step, even if we could: only
its existence is required for the contradiction.