Some cases in the threeway inversion lemma
The cases in this lemma are a little more verbose than those of previous
case analyses (in particular, the easiest way to discharge them seems to
be explicit manipulation on both each side goal and its matching
hypothesis). The direction of memory operations is likewise reversed,
and this is the more syntactically pedantic part of the new proof that
is not covered in previous ones.