Complete threeway inversion lemma
Still playing with existentials. The need for some variables to be in
scope by the time the logic variable for the top-level existential is
introduced forces us to anticipate some of the treatment for the side
goals up front, instead of leaving once the side goals have been
generated, which may look more natural.