Simplified memory initialization (first round)
Current memory initialization is a functional rendition of an imperative
process, which makes reasoning about the process rather involved.
However, the process lends itself to a cleaner and equivalent functional
treatment.

Further enhancements are possible once the equivalence of the substitute
has been adequately established, but these will be minor compared to the
present set of changes. New versions of the functions assume the names
of the old ones, which are currently primed for reference.