Move PC domain lemmas, rephrase silent star helper
As we suspected, the remaining helper required by the proof of
merge_states_silent_star is an elementary result, located a little later
in the development and in a slightly different form. Once the basic
lemmas are in scope, the result is a simple variation of one of those.

The proof of an existing lemma that uses one of the moved lemmas in
their section of origin needs to be tweaked to explicitly find the
assumptions that were previously implicit in section scope.