Move results on partial memories to Memory
Complete the library of existing results on partial memories in the
Memory module by relocating there the additions found elsewhere, mostly
in the intermediate PS.

Given their location there, it is natural to include the definitions for
partialization and combination of memories instead of employing
explicit and more cumbersom filterm operations.

At the moment, some derived results and notions of memory merging proper
to Recombination remain there.