Recombination cleanup (III)
Relocation of some lemmas, some minor restructuring, removal of a few
trivial (and unused) convenience rewrite lemmas which are no longer
used.

In the various sections, identify those that function as a "bag of
lemmas" and those that have a definite structure with one main theorem
and several supporting results.