Move unused files to outtakes folder
Separating those parts of the development which are not currently needed
to make it easy to measure the complexity of the system in its actual
state at any given time. One should be wary of outtakes bit rotting, but
they contain interesting results which required significant effort to
develop and may be useful in the future.