Complete round 1 of mergeable_states refactoring
In the final sections of the recomposition proof, we run into an
interweaving of lemmas that rely on external assumptions and lemmas
that are self-contained by way of the enriched definition of
mergeable_states, which need to be untangled to further advance the
simplification efforts.

In some parts, especially the need for external assumptions to
establish that the initial states are in the three-way simulation
relation, the changes highlight a limitation in our current approach:
because some functions like initial_state are always defined, regardless
of well-formedness, those conditions need to be imposed externally,
which is relatively inelegant vis-à-vis other related components.