Restructure three-way simulation sectioning
 - Group closely related results.

 - Remove redundant sections and merge compatible ones.

 - Separate sections reliant on external well-formedness assumptions
   from "self-contained" sections.

The proof is fairly well organized at the moment, with sections to
account for arguments of symmetry and assumptions required by its
various components.

Auxiliary results that are independent of the developments in this
part of the compiler are lodged in their corresponding files. Those
results that logically belong on this file are divided in two parts,
base results on mergeable states and simulation results, roughly equal
in extension.