Refine abstract memory model
In order to prove certain "relational" properties comparing pairs of
programs, their states (in particular, their initial) states need to be
inspected, and the initial memories of two "similar" programs should
reflect similarities in interfaces and code distribution.

This small extension to the abstract memory model, carrying over easily
to the concrete model used by the compiler, should be able to establish
at least a significant part of this connection (as discussed with
Jérémy).