Source PS becomes Blame
The use of parts of the old PS to prove the blame theorem through a
series of "partial execution" results should not obscure the object
of this part of the development: that is to say, the blame theorem.
Some remnants of the source partial semantics are retained for the
benefit of this proof, but they are not used anywhere else in the
proof and could conceivably be removed (albeit at a reasonably high
cost, by reengineering the proof to adopt a similar strategy to that
used, say, in the proof of recombination).