Canonical structure for extended events
Extend the embedding of events in (now nested) sum types and restore
the declaration of the canonical structure for this embedding. In doing
this, we follow FG's development, and declare a canonical structure for
the equality of values, which was not present in the master until now.

A question remains of whether to consider a pair of "undefined" value
constructors equal or not. Semantically, this may well not be the case,
but we may be getting ahead of ourselves by introducing this semantic
notion early on.

Co-authored-by: Akram El-Korashy <elkorashy@mpi-sws.org>