Structure missing case of genv_entrypoints lookup
Process the proof until we get to the core of the missing case. We
proceed by case analysis on one of the two analogous equalities; the two
sides can be easily kept in sync by appealing to program well-formedness
as an added hypothesis to the current formulation of the lemma -- as
usual, this is reasonable, and in the context of recombination, where
this result is needed, the requisite information is available in the
context.