Strengthen domm_prealloc
The strengthened (if still tentative) extensions to the abstract memory
model can be more informative than they currently are. For example, the
preallocation operation can not only impose that the number of allocated
blocks will coincide with the cardinality of the map of buffers passed
as an argument, but the domains of definition will be the same as well.
In fact, this is the only possibility that makes sense given the
interface of the abstract memory model.

It is expected that these properties will be of use in establishing the
connection between two program memories belonging to related programs.