Start changing the type path to be path_t.
path_t is the type of non-empty sequences of blocks, instead of just sequences of blocks.
The representation is inspired by ssreflect.path. The question of whether we should
directly use ssreflect.path is still unclear. My intuition is that we don't want to use
it because there (in ssreflect.path), path is defined merely as a boolean predicate.
Moreover, there is not lots of support for uniq---support for uniq would have been
the main appeal of using an existing path representation.