Coq 8.9 / 8.10 support; backwards compatibility
We aim to support newer Coq versions in combination with toolchains
composed of coeval versions of the libraries used by the development.
To this end, a natural choice is to favor combinations of libraries
supported by the official OPAM repositories.

In particular, versions of the Extensional Structures library map quite
cleanly to specific versions of Coq and Mathematical Structures. Most of
the development is unaffected by version changes, with a few exceptions:

 - In older toolchains, some case analyses on extensional structures are
   often followed by a rewrite of the corresponding case equality in the
   resulting sequent. In newer toolchains, this rewrite step is made
   implicitly as part of the [destruct] tactic. To support both older
   and newer toolchains, the existing rewrites must be made conditional.

 - The implicit parameters of certain SSReflect views, in particular
   [dommP], vary between older and newer toolchains. While newer
   toolchains admit more compact uses of said views, to support both
   older and newer toolchains simultaneously all implicit parameters
   must be applied explicitly using [@].

 - Most unpleasantly, due to changes in recent versions of SSReflect,
   the definitions of [elementsm] must differ across older and newer
   toolchains, although these changes are confined to the definition
   proper. The correct definition can be built conditionally in proof
   mode.

When newer versions of the toolchains are used, messages are emitted to
indicate the presence of backwards-compatible legacy code, which should
be removed once the older versions are no longer supported.