[QuickChick] Better scopes
Fix errors associated with interpretation scopes in recent versions of
QuickChick. With the exception of one case where decidable equality
comparison operators need to be disambiguated, all occurrences of this
kind of problem are related to the string scope and easily solved by
adding it to the scope stack in the affected test files.