[QuickChick] Better names in typeclass instances
Newer versions of QuickChick fix an issue present with some older
versions by which some names introduced automatically to the context
were unnecessarily renamed, yielding better maintainability. Those
instances of use of names that have not been explicitly declared are
corrected to reflect the original names given in the instance
declarations.

Since we are still in the process of adapting our tests to newer
toolchains as well as restoring the functionality that is known to have
been working in older toolchains, we preserve backwards compatibility by
means of conditional rewrites. When these are inactive, their presence
will be announced, so that they can be easily removed once support for
the older toolchains is dropped.