[QuickChick] Conditional show definition
The standard definition of [show_Z] is not available as such in older
versions of QuickChick, and needs to be replicated separately from
existing definitions, also in a way to signify that this is done
temporarily for backwards compatibility.

Puzzlingly, this adjustment is necessary in the version of the tools
for which testing was last reported as supported. It is possible that
some further changes were made after that point using newer tools, so
that the problem did not arise at that point.