[QuickChick] Test support for 8.8-9 toolchains
We can easily extend full test support to newer toolchains, in
particular those based on Coq 8.8 and 8.9 without any apparent changes
in test behavior (as noted in a previous commit, tests need to assess
their own behavior more carefully).

The most recent toolchain that we support, based on Coq 8.10, requires
some additional work. In particular, the extraction process generates
invalid OCaml code. This seems to be unrelated to the version of OCaml
being used. For example, 4.06.0 works with Coq 8.8 and 8.9 but not 8.10.