Add fingraph_of_mem_inv which, given a memory and invariants on this memory, creates a memory of finite nodes.
The memory of finite nodes can be used to instantiate all the lemmas from the fingraph library, most notably
the dfs_path_fin_mem.

TODO:
The function fingraph_of_mem_inv creates a memory of finite nodes. We need to prove
that this memory of finite nodes is equivalent to the original memories.

One difficulty now that prevents me from doing this proof is that I need to reason
about the helper function proofize_seq. In particular, I would like to prove the
lemma proofize_seq_spec. And I do not know how to do it. induction does not work
because it changes the dependent type in an undesired way.