Define cast functions that convert values of an ordinal type to values of a bigger ordinal type
Special credit of this idea goes to Jeremy and Adrien.

TODO: Change the way we finitize the memory. Instead of using the fact that the program is finite
and that the allocation status is finite, use a simpler fact (that is true of any memory), namely,
that by taking the maximum-pointer-value of all the pointers that *live* in the memory,
we are able to cast any pointer to a finite type defined by this maximum-pointer-value.

Credit of this second idea goes to Jeremy.

Co-authored-by: Catalin Hritcu <catalin.hritcu@gmail.com>
Co-authored-by: Rob Blanco <roberto.blanco@inria.fr>
Co-authored-by: Jérémy Thibault <jeremy.thibault@inria.fr>
Co-authored-by: Adrien Durier <adrien.durier@inria.fr>