[QuickChick] Test support for 8.10
Testing under Coq 8.10 works subject to the following adjustments. Due
to a bug introduced in the 8.10 branch of QuickChick, incorrect OCaml
code is generated. A workaround can be deployed by overwriting the
defective extraction directive before the [Extraction] commands are
executed these files:

 - `Tests/I2SFI/CompilerIntegrationTests.v`

 - `Tests/I2SFI/Test.v`

The command to be inserted is this:

  Extract Inductive Decimal.int => unit [ "(fun _ -> ())" "(fun _ -> ())" ] "(fun _ _ _ -> assert false)".

This solution does not work with Coq 8.7, since the library
`Coq.Init.Decimal` as only introduced in Coq 8.8.