Vacuously add new events to definability
To fix the remainder of the proof, all we need to do is fix
definability. This entails adding the new read and write events (with
arbitrary back-translations, as the events are not currently being
generated by the semantics) and discharging the new pair of (again,
currently nonsensical) goals generated over the course of the "prefix
lemma".

Of course, these will become real proofs as soon as we adapt the
semantics to emit the new events, so we need not perform any further
changes at the moment.

Co-authored-by: Akram El-Korashy <elkorashy@mpi-sws.org>
Co-authored-by: Jérémy Thibault <jeremy.thibault@inria.fr>