Initial integration of reachability in I.CS state
After adding the set of reachable addresses to program state in the
intermediate semantics, we need to propagate these changes to existing
code. In the semantics, temporarily and for the purpose of restoring the
build, we initialize with an empty set (a finite map, here), and
propagate it unchanged throughout the semantics (this is the expected
behavior in most cases).

Repairing the proofs is then a simple matter of amending case analyses
and adding missing variables where needed. Some of these proofs will
become non-trivial when the reachability set is updated by the
semantics, however its purely syntactic impact in the development is
small.

Co-authored-by: Adrien Durier <adrien.durier@inria.fr>
Co-authored-by: Akram El-Korashy <elkorashy@mpi-sws.org>
Co-authored-by: Jérémy Thibault <jeremy.thibault@inria.fr>