Add 'memory' to the non-informative event
Notice that this commit implements the following design decision:
The *whole* memory is emitted on the trace (at border states), and
the idea is that the trace relation will constrain only the
shared part of the memory. This design decision is crucial;
it allows us to avoid *computing* reachability and to instead
characterize the shared memory as a reachable part of the memory
that is given by just an inductive definition.

Notice also that as a result of emitting memory (whether the whole
memory or just the shared part) on the trace, the type of
trace events is now no longer an equality type.
This loss of decidable equality is not very catastrophic: after all,
we won't need it that much because the top-level theorems
and many lemmas will now be stated in terms of the trace relation
(that is not equality).

Some lemmas (like determinacy) will still need to be stated in terms
of decidable equality so that they remain conveniently usable.
(TODO: Check if it is possible to re-gain decidable equality in the
existence of memory on the trace event. Can we decide equality on
Memory.t?)