Patch up Definability
There is actually nothing wrong in the standard sequence of imports used
in Definability or in the shorthand modules S and I assembled in
S2I.Definitions: the semantics are visible in the latter, but not by
default in the individually imported constituents.

Restoring the definitions of those two modules, S and I, admitting a
couple of proofs in Blame and adding an auxiliary definition to lift
informative projections from traces to finite behavior prefixes lets us
avoid those problems.